Use floats with OpenCL.

+++ ksengine/kscontextcl.cpp
@@ -63,12 +63,12 @@ KSBuffer KSContextCL::createBuffer(const KSBuffer::BufferType  t,
                                    const Eigen::Matrix3Xd     &data)
 {
     cl_int err;
-    Matrix4Xd padded(4,data.cols());
-    padded.block(0,0,3,data.cols()) = data;
+    Matrix4Xf padded(4,data.cols());
+    padded.block(0,0,3,data.cols()) = data.cast<float>();
     void *bufdata = CAST_INTO_THE_VOID(padded.data());
     cl::Buffer clbuf(m_context,
     /* Type flags */ CL_MEM_READ_WRITE | CL_MEM_COPY_HOST_PTR,
-    /* # of bytes */ padded.size() * sizeof(double), 
+    /* # of bytes */ padded.size() * sizeof(float), 
     /* data ptr   */ bufdata, 
     /* error ptr  */ &err);
     if( err != CL_SUCCESS )
